refactor(component): Refactor loading states
