+++ apps/web/app/blog/layout.tsx
@@ -6,10 +6,5 @@ export const metadata = {
 export default function BlogLayout({
   children,
 }: Readonly<{ children: React.ReactNode }>) {
-  return (
-    <main className="min-h-screen p-8">
-      {/* <h2 className="mb-8 text-2xl font-bold tracking-tighter">Portfolio</h2> */}
-      {children}
-    </main>
-  );
+  return <main className="min-h-screen p-8">{children}</main>;
 }
